requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
streamlit
langchain
langchain-mistralai
langchain-core
plotly
pdfplumber
python-dotenv
pydantic
yfinance
pytz